refactor(frontend): decompose Accounts and Envelopes pages (861 -> ~470/500 lines each)
- components/accounts/: AccountTree (tree building + recursive AccountRow,
  moved out of the page) and AccountDialog — one form for create and edit,
  replacing the two near-verbatim dialog copies that differed only in ids,
  labels and parent options.
- components/envelopes/: EnvelopeTree and EnvelopeDialog with the same
  shape; the edit-only status field is driven by the mode. The
  manage-accounts dialog stays in the page (it has no create twin).
- Row components now use useTranslation directly instead of a threaded
  t prop.
